Refactor SpatioTemporalPath system: rename app from Malamute to BlackBear
- Updated all headers and object registrations to use BlackBearApp instead of MalamuteApp
- Replaced copyright/license headers accordingly
