+++ test/tests/spatiotemporal_path/tests
@@ -21,15 +21,4 @@
     design = 'ADMovingEllipsoidalHeatSource.md FunctionSpatioTemporalPath.md'
     requirement = 'The system shall be able to use a function-defined path to evolve the heat source.'
   []
-  [esm]
-    type = 'Exodiff'
-    input = 'esm.i'
-    exodiff = "esm_out.e
-               esm_out.e-s002 esm_out.e-s003 esm_out.e-s004
-               esm_out.e-s005 esm_out.e-s006 esm_out.e-s007
-               esm_out.e-s008 esm_out.e-s009"
-    design = 'SpatioTemporalPathElementSubdomainModifier.md'
-    mesh_mode = 'REPLICATED'
-    requirement = 'The system shall be able to use a SpatioTemporalPath to modify element subdomains.'
-  []
 []
